Update route.js
Great job with the project overall! The only comment I might have is cleaning up the code, getting rid of unneeded/commented out code, and also getting rid of console.logs in production. Professor suggested we add messages to show frontend when we encounter errors in the catch, so perhaps that would be a more user-centric design than using console.log() error.